Cadastro e recuperação só por e-mail

+++ prisma/migrations/20260203000000_add_professional_commission/migration.sql
@@ -1,2 +1,2 @@
--- Add commission percentage to Professional
-ALTER TABLE "Professional" ADD COLUMN "commissionPercentage" DOUBLE PRECISION NOT NULL DEFAULT 0;
+-- Add commission percentage to Professional (SQLite: REAL)
+ALTER TABLE "Professional" ADD COLUMN "commissionPercentage" REAL NOT NULL DEFAULT 0;
